requirements.txt
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.9.0
jsonschema==2.6.0